fast-api/requirements.txt
fastapi
uvicorn[standard]

python-dotenv
dropbox

joblib
numpy
pandas
scikit-learn
xgboost
